# perf: improve validation sync and async by replacing forEach with classic for loops