--- a/lib/document.js
+++ b/lib/document.js
@@ -2853,19 +2853,21 @@ Document.prototype.validateSync = function(pathsToValidate, options) {
   }
   const validating = {};
 
-  paths.forEach(function(path) {
+  for (let i = 0, len = paths.length; i < len; ++i) {
+    const path = paths[i];
+
     if (validating[path]) {
-      return;
+      continue;
     }
 
     validating[path] = true;
 
     const p = _this.$__schema.path(path);
     if (!p) {
-      return;
+      continue;
     }
     if (!_this.$isValid(path)) {
-      return;
+      continue;
     }
 
     const val = _this.$__getValue(path);
@@ -2879,11 +2881,11 @@ Document.prototype.validateSync = function(pathsToValidate, options) {
         p.$isArraySubdocument ||
         p.$isMongooseDocumentArray;
       if (isSubdoc && err instanceof ValidationError) {
-        return;
+        continue;
       }
       _this.invalidate(path, err, undefined, true);
     }
-  });
+  }
 
   const err = _this.$__.validationError;
   _this.$__.validationError = undefined;