--- a/lib/schematype.js
+++ b/lib/schematype.js
@@ -1250,10 +1250,10 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
     return fn(null);
   }
 
-  const _this = this;
-  validators.forEach(function(v) {
+  for (let i = 0, len = validators.length; i < len; ++i) {
+    const v = validators[i];
     if (err) {
-      return;
+      break;
     }
 
     const validator = v.validator;
@@ -1265,16 +1265,16 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
 
     if (validator instanceof RegExp) {
       validate(validator.test(value), validatorProperties);
-      return;
+      continue;
     }
 
     if (typeof validator !== 'function') {
-      return;
+      continue;
     }
 
-    if (value === undefined && validator !== _this.requiredValidator) {
+    if (value === undefined && validator !== this.requiredValidator) {
       validate(true, validatorProperties);
-      return;
+      continue;
     }
 
     try {
@@ -1303,8 +1303,7 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
     } else {
       validate(ok, validatorProperties);
     }
-
-  });
+  }
 
   function validate(ok, validatorProperties) {
     if (err) {