It is basically replacing the forEach with classic for loop

before
```
MONGOOSE_DEV=1 node validate
invalid x 23,482 ops/sec ±3.63% (89 runs sampled)
valid x 75,238 ops/sec ±0.58% (91 runs sampled)
```

after
```
MONGOOSE_DEV=1 node validate
invalid x 27,460 ops/sec ±0.91% (89 runs sampled)
valid x 76,128 ops/sec ±0.69% (93 runs sampled)
```